+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - `PulseUpdate::getUpdates()` now correctly returns an array of PulseUpdate objects
 - `PulseUpdate::getUpdates()` now checks it's parameters and attempts to automatically convert them to a format DaPulse will accept
 - `PulseUpdate::likeUpdate()` and `PulseUpdate::unlikeUpdate()` now returns boolean values on whether it was successful or not; it also throws an InvalidArgumentException when the given parameter is not a valid user
+- `Pulse::editPulseName()` has been added in order to allow for one less API call than creating a Pulse object and then calling `editName()`
 
 #### Deprecations